refactor: Remove DynamicValidations module and unused NovaSeq6000Validator
